Move lerp to math-utils
